requirements.txt

Django==2.0.1
psycopg2==2.7.5
absl-py==0.3.0
astor==0.7.1
gast==0.2.0
grpcio==1.13.0
h5py==2.8.0
setuptools==39.1.0
Keras==2.2.1
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
Markdown==2.6.11
numpy==1.15.0
protobuf==3.6.0
PyYAML==3.13
scipy==1.1.0
six==1.11.0
tensorboard==1.9.0
tensorflow==1.9.0
termcolor==1.1.0
Werkzeug==0.14.1

djangorestframework
django-filter
Pillow==5.0.0